+++ mayan/apps/storage/dependencies.py
@@ -1,6 +1,12 @@
 from mayan.apps.dependencies.classes import PythonDependency
 
 PythonDependency(
+    module=__name__, name='boto3', version_string='==1.24.70'
+)
+PythonDependency(
+    module=__name__, name='django-storages', version_string='==1.13.1'
+)
+PythonDependency(
     module=__name__, name='extract-msg', version_string='==0.34.3'
 )
 PythonDependency(